Fix reporting of period good frames
This should now update as per raw frames. There isn't a fast
appropriate call like total raw or good frames, but at least it
will now update. There is a wider issue to to this properly,
a method exists to return such information but only works if the
period card is running the periods so doesn't adjust for normal
software periods.